Fix argument normalization in fragments
Adds a initialization step to the argument normalization phase to fetch
the provided values from the Operation before traversing the blueprint
to normalize all inputs.

Current behavior is dependent on ordering of node traversal in the
blueprint i.e. Operation node is expected to be visited before Input
nodes. However, Fragment nodes are visited before the Operation node so
the provided values are empty.