Improve docs for global middleware
The middleware function has been updated to work on query, subscription
and mutation. These are likely to be the fields that users want
authentication on.

A note has also been added explaining why there is a fallback and what
the fallback does. This should discourage people adding the
Authentication middleware when it is not required (for example when
simple accessing a field on a map).