Add phase for setting context and root value
Some context values like the current logged in user-id are not available
during compilation phases of the document pipeline. So, add another
phase after all compilation phases to grab context values and put them in
blueprint execution context. This is needed when documents are
pre-compiled and are executed later when the context (like current
logged-in userID) values are actually available.

fixes #539