Fixes escape character sequences that are not on the last attributes (#590)
* Fixes escape character sequences that are not on the last attributes

When running with attributes that have escape characters and attributes
on the same line the order of the attributes matters for the outcome. If
the slashes appear in the last attribute then everything parses
successfully. If the attribute appears earlier in the list then an error
is returned instead.

The error:

~~~elixir
  validation_errors: [
    %Absinthe.Phase.Error{
      extra: %{},
      locations: [%{column: 0, line: 1}],
      message: "illegal: \") { name } }",
      path: [],
      phase: Absinthe.Phase.Parse
    }
  ]
~~~

This change allows the attributes to come in any order.

Amos King @adkron <amos@binarynoggin.com>

* Fix typo in test name