Add telemetry flag to Absinthe.run and resolution
See https://github.com/absinthe-graphql/absinthe/issues/807

The flag to Absinthe.run ensure the Telemetry phases are skipped in the
document pipeline.

Furthermore, the flag is passed on the
`Absinthe.Phase.Document.Execution.Resolution` phase where it sets
the (added) telemetry flag on the resolution struct. This is done so the
telemetry middleware can be skipped when the flag is set to false.