Ensure interface nullable types can be implemented by non-nullable types
Fixes https://github.com/absinthe-graphql/absinthe/issues/987

Two cases are added to the schema:
 * a simple nullable interface -> non-nullable implementor.
 * a nullable list -> with implementors `list_of(non_null(type))` and
   `non_null(list_of(non_null(type)))`.

From my reading of the spec these should all validate.

https://spec.graphql.org/draft/#IsValidImplementationFieldType()

I did find that it is not validated whether the object implements
the field's arguments. Also that the object does not implement
extra non-nullable arguments that are not specified in the interface,
this is not allowed. I'll create an issue for this.