copy logger metadata for async and batch
Currently whenever the async or batch middleware is used, any logger
metadata is lost when the new async tasks are created. This means that
the Phoenix request ID doesn't show in log lines if using Phoenix, and
it means that a user ID or other metadata needs to be added in every
resolver.

This copies the logger metadata from the parent process whenever a new
async task is created so that any logs further down the stack will have
access to it.