Store if document uses shorthand notation
It's not necessary for the execution of a document to preserve
this. For converting Absinthe.Language AST to a string it is however.

Note that it's not a goal that the document -> AST -> document
transformation is perfect in all cases. E.g. commas or lack thereof is
not preserved. The shorthand notation however seemed easy enough
to implement.

See https://spec.graphql.org/draft/#sec-Language.Operations.Query-shorthand