Convert `Absinthe.Language.*` structs to graphql representation (#1114)
* Store if document uses shorthand notation

It's not necessary for the execution of a document to preserve
this. For converting Absinthe.Language AST to a string it is however.

Note that it's not a goal that the document -> AST -> document
transformation is perfect in all cases. E.g. commas or lack thereof is
not preserved. The shorthand notation however seemed easy enough
to implement.

See https://spec.graphql.org/draft/#sec-Language.Operations.Query-shorthand

* Render Absinthe.Language AST to graphql string

* Move common utils to separate file

* Add Absinthe graphql formatter plugin

Note it doesn't implement the Mix.Tasks.Format behaviour since this
would give a warning pre Elixir 1.13

* Separate docs with two newlines