Extract deprecated directive fields into separate phase
This change was necessary to to implement macro-based directive
handling.

Because these fields are deprecated, they would use the new `deprecated`
directive. This directive definition is taken from the prototype schema.
However, during compilation of the prototype schema it would also invoke
the introspection builtins with 'deprecated' directive, which is not yet
available.

This new phase extracts the deprecated fields, thus allowing the
prototype phase to skip it in its schema pipeline to avoid the problem.

In Absinthe 2.0 the entire phase can be removed but as of now it's
a backwards incompatible change.