Ensure input/output types are correctly placed for SDL schemas
The InputOutputTypesCorrectlyPlaced phase did not work for
SDL schema's since it would unwrap a TypeReference, which
yields the TypeReference and not the type atom.

The subsequent lookup in the schema types would therefore fail, as
it expects atoms not typereferences, and it would not add errors.

It now correctly fetches the atom type identifier, though in a bit of
a roundabout way.