Support types with string identifiers
When generating a schema at runtime, it is desired that all dynamically generated identifiers be strings instead of atoms to prevent exhausting the `atom` table limit.

This does not change the macros `object` and `input_object` which still require types to be atoms, but when someone overrides the schema to provide runtime type identifiers, it resolves them properly.