Add telemetry events to async middleware
Introduce telemetry events to Async middleware to instrument
new `Task.async/1` created inside the middleware.

The use-case for this event comes from OpenTelemetry and trace
propagation. For OTel, we need to run instrumentation from inside
the newly created Task to be able to connect spans. With this event
in place, we can track those Tasks without introducing custom wrappers
on user code.

This patch adds events only in the case of the middleware creating the
task. For tasks created by user code, there is nothing the library can
do, unfortunately.