feat: ability to run custom callback on timeout for Absinthe.Middleware.Async middleware
Async middleware sends an exit signal that kills the request and causes a 500 on timeout, as it uses
Task.await/2 under the hood. This PR adds the ability to pass on_timeout/1 callback function to
the middleware and trigger it on timeout.
By default, it stops the process current process with exit/1, as before.

```elixir
field :time_consuming, :thing do
  resolve fn _, _, _ ->
    async(
      fn -> {:ok, long_time_consuming_function()} end,
      timeout: 7_000,
      on_timeout: fn _timeout ->
        {:error, "Failed to get a result"}
      end
    )
  end
end
```