Ensure root objects are marked as referenced correctly
Followup to https://github.com/absinthe-graphql/absinthe/pull/1176

Since every schema will have a declared or inferred declaration, this
can be used to mark root objects as referenced.

The MarkReferenced used to mark every :query/:subscription/:mutation object
as referenced, even when it was used in the declaration.

This fix ensures that root objects that are not used in the declaration
are not marked as referenced and thus not rendered in the SDL.