Fix default enum value present for SDL schemas
TypeReference.unwrap/1 was passed in %TypeReference.Name{} struct in the
case of SDL schema's, and returns it. This fails in the subsequent lookup
in `enums`. We need to use TypeReference.to_type/2 to get
the actual atom type identifier for the `enums` lookup.

For macro schema's this is not an issue as the atom type identifier is
passed into TypeReference.unwrap/1.