Validate type references for invalid wrapped types
In some places we were not checking for invalid wrapped types
(non-null/list). E.g.
```elixir
union :my_union do
  types [:boolean, list_of(:id)]
end
```

Or when using `import_fields` or `interface` with a wrapped type.

This PR adds the missing checks in the TypeReferencesExist phase. All
cases where a wrapped type is invalid are now caught in this phase.

What still needs to be tackled is whether the types passed in are valid
for that case. A union's types can only consist of object types, interfaces
can only refer to interface definitions etc. This could be done in a
separate phase. Some of this work is (incidentally) tackled by other
phases but I think it makes sense to group this.

Fixes #1048