Fix SDL type definition rendering order
In OTP-25 the SDL rendering was different compared to earlier versions.

I traced it down to `:digraph_utils.topsort()` although could not
find a relevant change in OTP-25 that would explain the difference.

`topsort/1` is used to sort the nodes in the graph so that the import
process can be done in the correct order in the FieldImports phase.
I've changed it to no longer rely on `topsort/1`. Since the
`NoCircularFieldImports` ensures there are no cycles in the graph
we can use recursion for solving the imports without ending up in an
endless loop.

This could be less efficient. The `topsort/1` ensured that the order
of imports was such that when `A` imported `B`, `B` was already imported,
etc up the hierachy chain. If `C` also imported `B`, the imports for
`B` would also already be done. So, `import_fields(B)` would be called once
for the entire graph.

Using recursion, `import_fields(B)` would be called twice, once for `A`
and once for `C`. If B itself imported other fields, those would also be
called multiple times.

I don't think this will be a problem in practice. The cost is only
incurred for schema building, and the `import_fields` graphs are
probably pretty shallow.

The rendering order now matches the order the type definitions were
defined in.