Allow `Registry` compression to be configured by users
At the moment `Registry` is started with `compressed: true` hard coded.

This presented problems when we updated absinthe some time ago, we are
running an application with a large number of subscriptions and so the
added CPU utilization was quite large. To avoid having to scale up we
pinned absinthe to a pre-compression version.

We'd like to be able to keep up to date, and so propose this change to
allow users to opt-out of compression and would welcome your feedback.

This PR alters the type signature of
`Absinthe.Subscription.child_spec/1` so users can pass more
configuration options. The original argument (a single pub-sub) is still
supported for backwards compatibility.