Allow configurable docset runner for subscription documents
As the docs mention, currently when subscription documents are run
it provides a simple form of back pressure due to it blocking the mutation.
Furthermore there is no batching across subscriptions.

This PR gives the ability to configure a docset runner by implementing
the Absinthe.Subscription.DocSetRunner behaviour.

A custom runner could be implemented to provide batching,
or to provide a different form of back pressure. Both of these
are useful for scaling out subscriptions. They can also both be highly
specific to the project and so it makes sense to allow them to be configured.
Furthermore, it also gives the opportunity to adjust the pipeline
for subscription documents just before execution.

I've changed the docset runner to accept a struct instead of a tuple,
as it makes it easier to add new options in the future.

The default is the same as before, a simple runner that runs the docset
and blocks the mutation.