Bail from subscription if resolver has an error
Would it be possible to do something like this? Right now, null is sent down to the subscriber if the resolver produces an error. 

It might depend on the use case, but I was thinking about something where you have a table where users can only see some of the entries and you have a subscription to updates, for example and use the user_id as the context id. Now, in the resolver, you check if the user is authorized to see the updated value and if not, put an error into the resolution. With the current implementation, the user would still get a push from the subscription with the null value. But not notifying the user in this case might be the better thing to do.