fix: use `Code.ensure_compiled!` at compile time
While investigating the changes in this PR:
https://github.com/elixir-lang/elixir/pull/14451

I encountered an error about not being able to load a module at
this line, implying this is called at compile time.

@josevalim pointed out that `Code.ensure_loaded!/1` will have no effect
at compile time, and so this was really just relying on compile order
and would be non-deterministic. `Code.ensure_compiled!/1` is a superset
of `Code.ensure_loaded!/1` and will work at compile time to ensure the depended
on module is compiled.

This would appear to solve #1351.