Fix error return when the value is a struct
When returning a struct implementing to_string an exception will be
thrown on Enum.to_list, because structs passes the is_map/1 gaurd and
Enum.to_list fails on structs. In these cases it needs to fall back to to_string.

In Elixir 1.17 we can use is_non_struct_map/1, but think it won't be
worth updating the minimum Elixir version for this.